move all commands to separate files
