--- src/pipx/commands/commands.py
@@ -1,406 +1,0 @@
-"""The implementation of pipx commands"""
-
-import logging
-import sys
-import textwrap
-import tempfile
-import time
-from pathlib import Path
-from shutil import which
-from typing import List, Optional
-
-import userpath  # type: ignore
-from pipx import constants
-from pipx.colors import bold
-from pipx.commands.common import expose_apps_globally, get_package_summary
-from pipx.emojies import hazard, stars
-from pipx.util import WINDOWS, PipxError, rmdir, valid_pypi_name
-from pipx.venv import Venv, VenvContainer
-
-
-def install(
-    venv_dir: Optional[Path],
-    package_name: Optional[str],
-    package_spec: str,
-    local_bin_dir: Path,
-    python: str,
-    pip_args: List[str],
-    venv_args: List[str],
-    verbose: bool,
-    *,
-    force: bool,
-    include_dependencies: bool,
-):
-    # package_spec is anything pip-installable, including package_name, vcs spec,
-    #   zip file, or tar.gz file.
-    # Determine package_name to properly name venv directory.
-    if venv_dir is None or package_name is None:
-        package_name = _package_name_from_spec(
-            package_spec, python, pip_args=pip_args, verbose=verbose
-        )
-        venv_container = VenvContainer(constants.PIPX_LOCAL_VENVS)
-        venv_dir = venv_container.get_venv_dir(package_name)
-
-    try:
-        exists = venv_dir.exists() and next(venv_dir.iterdir())
-    except StopIteration:
-        exists = False
-
-    if exists:
-        if force:
-            print(f"Installing to existing directory {str(venv_dir)!r}")
-        else:
-            print(
-                f"{package_name!r} already seems to be installed. "
-                f"Not modifying existing installation in {str(venv_dir)!r}. "
-                "Pass '--force' to force installation."
-            )
-            return
-
-    venv = Venv(venv_dir, python=python, verbose=verbose)
-    try:
-        venv.create_venv(venv_args, pip_args)
-        venv.install_package(
-            package=package_name,
-            package_or_url=package_spec,
-            pip_args=pip_args,
-            include_dependencies=include_dependencies,
-            include_apps=True,
-            is_main_package=True,
-        )
-        _run_post_install_actions(
-            venv,
-            package_name,
-            local_bin_dir,
-            venv_dir,
-            include_dependencies,
-            force=force,
-        )
-    except (Exception, KeyboardInterrupt):
-        print("")
-        venv.remove_venv()
-        raise
-
-
-def _package_name_from_spec(
-    package_spec: str, python: str, *, pip_args: List[str], verbose: bool
-) -> str:
-    start_time = time.time()
-
-    # shortcut if valid PyPI name and not a local path
-    if valid_pypi_name(package_spec) and not Path(package_spec).exists():
-        package_name = package_spec
-        logging.info(f"Determined package name: {package_name}")
-        logging.info(f"Package name determined in {time.time()-start_time:.1f}s")
-        return package_name
-
-    with tempfile.TemporaryDirectory() as temp_venv_dir:
-        venv = Venv(Path(temp_venv_dir), python=python, verbose=verbose)
-        venv.create_venv(venv_args=[], pip_args=[])
-        package_name = venv.install_package_no_deps(
-            package_or_url=package_spec, pip_args=pip_args
-        )
-
-    logging.info(f"Package name determined in {time.time()-start_time:.1f}s")
-    return package_name
-
-
-def _run_post_install_actions(
-    venv: Venv,
-    package: str,
-    local_bin_dir: Path,
-    venv_dir: Path,
-    include_dependencies: bool,
-    *,
-    force: bool,
-):
-    package_metadata = venv.package_metadata[package]
-
-    if not package_metadata.app_paths and not include_dependencies:
-        # No apps associated with this package and we aren't including dependencies.
-        # This package has nothing for pipx to use, so this is an error.
-        for dep, dependent_apps in package_metadata.app_paths_of_dependencies.items():
-            print(
-                f"Note: Dependent package '{dep}' contains {len(dependent_apps)} apps"
-            )
-            for app in dependent_apps:
-                print(f"  - {app.name}")
-
-        if venv.safe_to_remove():
-            venv.remove_venv()
-
-        if len(package_metadata.app_paths_of_dependencies.keys()):
-            raise PipxError(
-                f"No apps associated with package {package}. "
-                "Try again with '--include-deps' to include apps of dependent packages, "
-                "which are listed above. "
-                "If you are attempting to install a library, pipx should not be used. "
-                "Consider using pip or a similar tool instead."
-            )
-        else:
-            raise PipxError(
-                f"No apps associated with package {package}. "
-                "If you are attempting to install a library, pipx should not be used. "
-                "Consider using pip or a similar tool instead."
-            )
-
-    if package_metadata.apps:
-        pass
-    elif package_metadata.apps_of_dependencies and include_dependencies:
-        pass
-    else:
-        # No apps associated with this package and we aren't including dependencies.
-        # This package has nothing for pipx to use, so this is an error.
-        if venv.safe_to_remove():
-            venv.remove_venv()
-        raise PipxError(
-            f"No apps associated with package {package} or its dependencies."
-            "If you are attempting to install a library, pipx should not be used. "
-            "Consider using pip or a similar tool instead."
-        )
-
-    expose_apps_globally(
-        local_bin_dir, package_metadata.app_paths, package, force=force
-    )
-
-    if include_dependencies:
-        for _, app_paths in package_metadata.app_paths_of_dependencies.items():
-            expose_apps_globally(local_bin_dir, app_paths, package, force=force)
-
-    print(get_package_summary(venv_dir, package=package, new_install=True))
-    _warn_if_not_on_path(local_bin_dir)
-    print(f"done! {stars}", file=sys.stderr)
-
-
-def _warn_if_not_on_path(local_bin_dir: Path):
-    if not userpath.in_current_path(str(local_bin_dir)):
-        logging.warning(
-            f"{hazard}  Note: {str(local_bin_dir)!r} is not on your PATH environment "
-            "variable. These apps will not be globally accessible until "
-            "your PATH is updated. Run `pipx ensurepath` to "
-            "automatically add it, or manually modify your PATH in your shell's "
-            "config file (i.e. ~/.bashrc)."
-        )
-
-
-def inject(
-    venv_dir: Path,
-    package_name: Optional[str],
-    package_spec: str,
-    pip_args: List[str],
-    *,
-    verbose: bool,
-    include_apps: bool,
-    include_dependencies: bool,
-    force: bool,
-):
-    if not venv_dir.exists() or not next(venv_dir.iterdir()):
-        raise PipxError(
-            textwrap.dedent(
-                f"""\
-            Can't inject {package_spec!r} into nonexistent Virtual Environment {str(venv_dir)!r}.
-            Be sure to install the package first with pipx install {venv_dir.name!r}
-            before injecting into it."""
-            )
-        )
-
-    venv = Venv(venv_dir, verbose=verbose)
-
-    # package_spec is anything pip-installable, including package_name, vcs spec,
-    #   zip file, or tar.gz file.
-    if package_name is None:
-        package_name = _package_name_from_spec(
-            package_spec, venv.python, pip_args=pip_args, verbose=verbose
-        )
-
-    venv.install_package(
-        package=package_name,
-        package_or_url=package_spec,
-        pip_args=pip_args,
-        include_dependencies=include_dependencies,
-        include_apps=include_apps,
-        is_main_package=False,
-    )
-    if include_apps:
-        _run_post_install_actions(
-            venv,
-            package_name,
-            constants.LOCAL_BIN_DIR,
-            venv_dir,
-            include_dependencies,
-            force=force,
-        )
-
-    print(f"  injected package {bold(package_name)} into venv {bold(venv_dir.name)}")
-    print(f"done! {stars}", file=sys.stderr)
-
-
-def uninstall(venv_dir: Path, package: str, local_bin_dir: Path, verbose: bool):
-    """Uninstall entire venv_dir, including main package and all injected
-    packages.
-    """
-    if not venv_dir.exists():
-        print(f"Nothing to uninstall for {package} 😴")
-        app = which(package)
-        if app:
-            print(
-                f"{hazard}  Note: '{app}' still exists on your system and is on your PATH"
-            )
-        return
-
-    venv = Venv(venv_dir, verbose=verbose)
-
-    if venv.pipx_metadata.main_package is not None:
-        app_paths: List[Path] = []
-        for viewed_package in venv.package_metadata.values():
-            app_paths += viewed_package.app_paths
-            for dep_paths in viewed_package.app_paths_of_dependencies.values():
-                app_paths += dep_paths
-    else:
-        # fallback if not metadata from pipx_metadata.json
-        if venv.python_path.is_file():
-            # has a valid python interpreter and can get metadata about the package
-            metadata = venv.get_venv_metadata_for_package(package)
-            app_paths = metadata.app_paths
-            for dep_paths in metadata.app_paths_of_dependencies.values():
-                app_paths += dep_paths
-        else:
-            # Doesn't have a valid python interpreter. We'll take our best guess on what to uninstall
-            # here based on symlink location. pipx doesn't use symlinks on windows, so this is for
-            # non-windows only.
-            # The heuristic here is any symlink in ~/.local/bin pointing to .local/pipx/venvs/PACKAGE/bin
-            # should be uninstalled.
-            if WINDOWS:
-                app_paths = []
-            else:
-                apps_linking_to_venv_bin_dir = [
-                    f
-                    for f in constants.LOCAL_BIN_DIR.iterdir()
-                    if str(f.resolve()).startswith(str(venv.bin_path))
-                ]
-                app_paths = apps_linking_to_venv_bin_dir
-
-    for file in local_bin_dir.iterdir():
-        if WINDOWS:
-            for b in app_paths:
-                if file.name == b.name:
-                    file.unlink()
-        else:
-            symlink = file
-            for b in app_paths:
-                if symlink.exists() and b.exists() and symlink.samefile(b):
-                    logging.info(f"removing symlink {str(symlink)}")
-                    symlink.unlink()
-
-    rmdir(venv_dir)
-    print(f"uninstalled {package}! {stars}")
-
-
-def uninstall_all(venv_container: VenvContainer, local_bin_dir: Path, verbose: bool):
-    for venv_dir in venv_container.iter_venv_dirs():
-        package = venv_dir.name
-        uninstall(venv_dir, package, local_bin_dir, verbose)
-
-
-def reinstall_all(
-    venv_container: VenvContainer,
-    local_bin_dir: Path,
-    python: str,
-    verbose: bool,
-    *,
-    skip: List[str],
-):
-    for venv_dir in venv_container.iter_venv_dirs():
-        package = venv_dir.name
-        if package in skip:
-            continue
-
-        venv = Venv(venv_dir, verbose=verbose)
-
-        if venv.pipx_metadata.main_package.package_or_url is not None:
-            package_or_url = venv.pipx_metadata.main_package.package_or_url
-        else:
-            package_or_url = package
-
-        uninstall(venv_dir, package, local_bin_dir, verbose)
-
-        # install main package first
-        install(
-            venv_dir,
-            package,
-            package_or_url,
-            local_bin_dir,
-            python,
-            venv.pipx_metadata.main_package.pip_args,
-            venv.pipx_metadata.venv_args,
-            verbose,
-            force=True,
-            include_dependencies=venv.pipx_metadata.main_package.include_dependencies,
-        )
-
-        # now install injected packages
-        for (
-            injected_name,
-            injected_package,
-        ) in venv.pipx_metadata.injected_packages.items():
-            if injected_package.package_or_url is None:
-                # This should never happen, but package_or_url is type
-                #   Optional[str] so mypy thinks it could be None
-                raise PipxError("Internal Error injecting package")
-            inject(
-                venv_dir,
-                injected_name,
-                injected_package.package_or_url,
-                injected_package.pip_args,
-                verbose=verbose,
-                include_apps=injected_package.include_apps,
-                include_dependencies=injected_package.include_dependencies,
-                force=True,
-            )
-
-
-def run_pip(package: str, venv_dir: Path, pip_args: List[str], verbose: bool):
-    venv = Venv(venv_dir, verbose=verbose)
-    if not venv.python_path.exists():
-        raise PipxError(
-            f"venv for {package!r} was not found. Was {package!r} installed with pipx?"
-        )
-    venv.verbose = True
-    venv._run_pip(pip_args)
-
-
-def ensurepath(location: Path, *, force: bool):
-    location_str = str(location)
-
-    post_install_message = (
-        "You likely need to open a new terminal or re-login for "
-        "the changes to take effect."
-    )
-    if userpath.in_current_path(location_str) or userpath.need_shell_restart(
-        location_str
-    ):
-        if not force:
-            if userpath.need_shell_restart(location_str):
-                print(
-                    f"{location_str} has been already been added to PATH. "
-                    f"{post_install_message}"
-                )
-            else:
-                logging.warning(
-                    (
-                        f"The directory `{location_str}` is already in PATH. If you "
-                        "are sure you want to proceed, try again with "
-                        "the '--force' flag.\n\n"
-                        f"Otherwise pipx is ready to go! {stars}"
-                    )
-                )
-            return
-
-    userpath.append(location_str)
-    print(f"Success! Added {location_str} to the PATH environment variable.")
-    print(
-        "Consider adding shell completions for pipx. "
-        "Run 'pipx completions' for instructions."
-    )
-    print()
-    print(f"{post_install_message} {stars}")
