+++ src/pipx/commands/uninstall_all.py
@@ -1,0 +1,10 @@
+from pathlib import Path
+
+from pipx.venv import VenvContainer
+from pipx.commands import uninstall
+
+
+def uninstall_all(venv_container: VenvContainer, local_bin_dir: Path, verbose: bool):
+    for venv_dir in venv_container.iter_venv_dirs():
+        package = venv_dir.name
+        uninstall(venv_dir, package, local_bin_dir, verbose)
